Add files via upload
Write a program that includes a function that takes three int type array input arguments
and their corresponding size, and sum the corresponding elements of the first two arrays,
and put the results into the third array. In your program make sure that you get the size of
the arrays from the user and creates the arrays in the main, make use of pointers and
dynamic memory allocation for this.